Revert "[Issue NexClipper#101] Change Toggle Button design and Rearrange it"
